cli(topic): reject zero values for --count / --duration
Review feedback on #235. Without a range check, '--count 0' and
'--duration 0' silently no-op'd and returned empty output — a typo
would look like success in scripting/CI contexts.

Now enforced via clap's built-in range validator (range(1..)), matching
the pre-existing 'parse_window' pattern used by --window. Invalid
values fail at argument parse time with exit code 2 and a clear error:

  error: invalid value '0' for '--count <N>': 0 is not in 1..<max>

Also bumped the echo --count type from usize to u64 so clap's
value_parser!(u64).range(1..) applies uniformly on all platforms
(usize has different widths across architectures).

+++ binaries/cli/src/command/topic/echo.rs
@@ -49,12 +49,14 @@ pub struct Echo {
     #[clap(long, value_name = "FORMAT", default_value_t = OutputFormat::Table)]
     pub format: OutputFormat,
 
-    /// Exit after this many messages (default: stream until interrupted)
-    #[clap(long, value_name = "N")]
-    pub count: Option<usize>,
+    /// Exit after this many messages (default: stream until interrupted).
+    /// Must be at least 1.
+    #[clap(long, value_name = "N", value_parser = clap::value_parser!(u64).range(1..))]
+    pub count: Option<u64>,
 
-    /// Exit after this many seconds (default: stream until interrupted)
-    #[clap(long, value_name = "SECONDS")]
+    /// Exit after this many seconds (default: stream until interrupted).
+    /// Must be at least 1.
+    #[clap(long, value_name = "SECONDS", value_parser = clap::value_parser!(u64).range(1..))]
     pub duration: Option<u64>,
 
     #[clap(flatten)]
@@ -79,7 +81,7 @@ fn inspect(
     coordinator: CoordinatorOptions,
     selector: TopicSelector,
     format: OutputFormat,
-    count: Option<usize>,
+    count: Option<u64>,
     duration: Option<u64>,
 ) -> eyre::Result<()> {
     let session = coordinator.connect()?;
@@ -96,7 +98,7 @@ fn inspect(
     const HINT_TIMEOUT: std::time::Duration = std::time::Duration::from_secs(5);
     let mut hint_shown = false;
     let mut buf = Vec::with_capacity(1024);
-    let mut emitted: usize = 0;
+    let mut emitted: u64 = 0;
     let deadline = duration.map(|s| std::time::Instant::now() + std::time::Duration::from_secs(s));
     loop {
         // Stop conditions: --count reached or --duration elapsed.
